requirements.txt
anyio==4.12.1
blinker==1.9.0
click==8.3.1
cloudevents==1.12.0
deprecation==2.1.0
Flask==3.1.2
functions-framework==3.10.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.2.1
starlette==0.51.0
typing_extensions==4.15.0
uvicorn==0.40.0
uvicorn-worker==0.4.0
watchdog==6.0.0
Werkzeug==3.1.5
